# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/employee/{id}": {
            "delete": {
                "description": "get employee by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a employee entity by ID",
                "operationId": "delete-employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/employees": {
            "get": {
                "description": "list employee entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List employee entities",
                "operationId": "list-employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Employee"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create employee",
                "operationId": "create-employee",
                "parameters": [
                    {
                        "description": "Employee entity",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Employee"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/employees/{id}": {
            "get": {
                "description": "get employee by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a employee entity by ID",
                "operationId": "get-employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Employee"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/faculty/{id}": {
            "delete": {
                "description": "get faculty by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a faculty entity by ID",
                "operationId": "delete-faculty",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Faculty ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/facultys": {
            "get": {
                "description": "list faculty entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List faculty entities",
                "operationId": "list-faculty",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Faculty"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create faculty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create faculty",
                "operationId": "create-faculty",
                "parameters": [
                    {
                        "description": "Faculty entity",
                        "name": "faculty",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Faculty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Faculty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/facultys/{id}": {
            "get": {
                "description": "get faculty by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a faculty entity by ID",
                "operationId": "get-faculty",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Faculty ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Faculty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Branch": {
            "type": "object",
            "properties": {
                "brname": {
                    "description": "Brname holds the value of the \"brname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BranchQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BranchEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BranchEdges": {
            "type": "object",
            "properties": {
                "faculty": {
                    "description": "Faculty holds the value of the faculty edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Faculty"
                },
                "userInformations": {
                    "description": "UserInformations holds the value of the user_informations edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.User"
                    }
                }
            }
        },
        "ent.Building": {
            "type": "object",
            "properties": {
                "buname": {
                    "description": "Buname holds the value of the \"buname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BuildingQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BuildingEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Employee": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EmployeeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.EmployeeEdges"
                },
                "employeeemail": {
                    "description": "Employeeemail holds the value of the \"employeeemail\" field.",
                    "type": "string"
                },
                "employeename": {
                    "description": "Employeename holds the value of the \"employeename\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.EmployeeEdges": {
            "type": "object",
            "properties": {
                "employees": {
                    "description": "Employees holds the value of the employees edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Returninvoice"
                    }
                }
            }
        },
        "ent.Faculty": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the FacultyQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.FacultyEdges"
                },
                "fname": {
                    "description": "Fname holds the value of the \"fname\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.FacultyEdges": {
            "type": "object",
            "properties": {
                "branchs": {
                    "description": "Branchs holds the value of the branchs edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Branch"
                    }
                },
                "userInformations": {
                    "description": "UserInformations holds the value of the user_informations edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.User"
                    }
                }
            }
        },
        "ent.Repairinvoice": {
            "type": "object",
            "properties": {
                "deviceid": {
                    "description": "Deviceid holds the value of the \"deviceid\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RepairinvoiceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RepairinvoiceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "statusrepairid": {
                    "description": "Statusrepairid holds the value of the \"statusrepairid\" field.",
                    "type": "integer"
                },
                "symptomid": {
                    "description": "Symptomid holds the value of the \"symptomid\" field.",
                    "type": "integer"
                },
                "userid": {
                    "description": "Userid holds the value of the \"userid\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.RepairinvoiceEdges": {
            "type": "object",
            "properties": {
                "repairinvoices": {
                    "description": "Repairinvoices holds the value of the repairinvoices edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Returninvoice"
                }
            }
        },
        "ent.Returninvoice": {
            "type": "object",
            "properties": {
                "addedtime": {
                    "description": "Addedtime holds the value of the \"addedtime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ReturninvoiceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ReturninvoiceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ReturninvoiceEdges": {
            "type": "object",
            "properties": {
                "employee": {
                    "description": "Employee holds the value of the Employee edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Employee"
                },
                "repairinvoice": {
                    "description": "Repairinvoice holds the value of the Repairinvoice edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Repairinvoice"
                },
                "statust": {
                    "description": "Statust holds the value of the Statust edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Statust"
                }
            }
        },
        "ent.Room": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "rname": {
                    "description": "Rname holds the value of the \"rname\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Statust": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the StatustQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StatustEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "statustname": {
                    "description": "Statustname holds the value of the \"statustname\" field.",
                    "type": "string"
                }
            }
        },
        "ent.StatustEdges": {
            "type": "object",
            "properties": {
                "statusts": {
                    "description": "Statusts holds the value of the statusts edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Returninvoice"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "personalID": {
                    "description": "PersonalID holds the value of the \"personalID\" field.",
                    "type": "string"
                },
                "personalName": {
                    "description": "PersonalName holds the value of the \"personalName\" field.",
                    "type": "string"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "branch": {
                    "description": "Branch holds the value of the branch edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Branch"
                },
                "building": {
                    "description": "Building holds the value of the building edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Building"
                },
                "faculty": {
                    "description": "Faculty holds the value of the faculty edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Faculty"
                },
                "room": {
                    "description": "Room holds the value of the room edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Room"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}